cls_rbd: use the require_feature function to make sure we have new-style headers
Previously, some functions weren't checking for new-style headers at all,
and lots of the ones that did check were doing so incidentally to
looking up a value. This meant that if you tried to do new ops on an
old header, you might succeed, you might get an ENOENT, or you might
fail. Now, we check at the beginning of these functions and return
-ENOEXEC if it's not a new-style header.

Signed-off-by: Greg Farnum <greg@inktank.com>