mon: Paxos: modify the criteria to propose queued proposals in finish_proposal()
This is nothing much. We just propose everytime we finish a proposal, as
long as we are in state ACTIVE and the head of the proposal queue has not
been proposed yet.

Also, we cannot assume the head of the queue has been proposed, since this
function is called from message handler functions, and we incur in the
possibility of being called more than once during the same proposal (fyi,
this usually happens from handle_accept(), which leads us to believe
that there is the possibility of a same message to be delivered twice).

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>