mon: PGMonitor: Wait for readable and writable using PaxosService interface
We cannot wait on Paxos if we are waiting for PaxosService state changes.
For instance, the service may not be writable because it has a queued
proposal, but the Paxos may be (even though momentarily) on state ACTIVE.

We need to cope with this, and the way to do it is by using the PaxosService
wait_for_*() functions.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>