mon: AuthMonitor: wait for finished proposal instead of commit
We have been waiting for Paxos to finish committing in order to send our
command ack messages. Although this does not pose a problem per se, since
it is assumed that it should *eventually* be true, there is a chance that
it might not. It happens that if we bootstrap before our proposal is
handled, our proposal will not be taken care of and acking our command is
and invalid behavior. Thus, we really must depend solely on the PaxosService
waiting queue for finer grained control.

Let me know if this isn't as clear as possible.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>