mon: PaxosService: handle the case when first_committed_version equals 0
In a nutshell, the services do not set the first committed version, as they
consider it as a SEP (Somebody Else's Problem). They do rely on it though,
and we, the PaxosService, must ensure that it contains a valid value (that
is, higher than zero) at all times.

Since we will only have a first_committed version equal to zero once, and
that is before the service's first proposal, we are safe to simply read
the variable from the store and assign the first_committed the same value
as the last_committed iff the first_committed version is zero.

This also affects trimming, since trimming relies on the first_committed
version as the lower bound for version trimming. Even though the k/v store
will gracefully ignore any problem from trying to remove non-existent
versions, the main issue would still stand: we'd be removing a non-existent
version and that just doesn't make any sense.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>